Rename: security 구조 변경
