+++ src/test/java/org/example/spring/security/CustomUsernamePasswordAuthenticationProviderTest.java
@@ -7,6 +7,8 @@ import static org.junit.jupiter.api.Assertions.assertTrue;
 import static org.mockito.Mockito.when;
 
 import java.util.Collections;
+import org.example.spring.security.service.CustomUserDetailsService;
+import org.example.spring.security.service.CustomUsernamePasswordAuthenticationProvider;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
